+++ bcbio/chipseq/peaks.py
@@ -14,7 +14,7 @@ def get_callers():
     from bcbio.chipseq import macs2
     return {"macs2": macs2.run}
 
-def peakcall_preparation(data, run_parallel):
+def peakcall_prepare(data, run_parallel):
     """Entry point for doing peak calling"""
     caller_fns = get_callers()
     to_process = []
@@ -28,7 +28,7 @@ def peakcall_preparation(data, run_parallel):
                 if mimic:
                     to_process.append(mimic)
                 else:
-                    logger.info("No input sample for %s" % name)
+                    logger.info("Skipping peak calling. No input sample for %s" % name)
     if to_process:
         after_process = run_parallel("peakcalling", to_process)
         data = _sync(data, after_process)
@@ -42,7 +42,6 @@ def calling(data):
     name = dd.get_sample_name(data)
     out_dir = utils.safe_makedir(os.path.join(dd.get_work_dir(data), data["peak_fn"], name ))
     out_file = caller_fn(name, chip_bam, input_bam, dd.get_genome_build(data), out_dir, data["config"])
-# utils.symlink_plus(call_file, out_file)
     data["peaks_file"] = out_file
     return [[data]]
 
@@ -53,10 +52,11 @@ def _sync(original, processed):
     to consider multiple callers.
     """
     for original_sample in original:
+        original_sample[0]["peaks_file"] = []
         for processs_sample in processed:
             if dd.get_sample_name(original_sample[0]) == dd.get_sample_name(processs_sample[0]):
                 if processs_sample[0]["peaks_file"] != "error":
-                    original_sample[0]["peaks_file"] = processs_sample[0]["peaks_file"]
+                    original_sample[0]["peaks_file"].append(processs_sample[0]["peaks_file"])
     return original
 
 def _get_paired_samples(sample, data):